Create tab panel behaviour
